Update car_finder2.py
Changes made to this version:
Only scrapes the data of cars not already scraped. Urls of all scraped cars are stored to binary file. Per urls, it first sees if a url was already scraped by referring to the binary file. 
All new items are directly added to a csv file. Old version created a dataframe per brand and then wrote it to file.